+++ src/clients/emrServerlessClient.ts
@@ -4,10 +4,14 @@
 import * as vscode from "vscode";
 import {
   EMRServerlessClient,
+  JobDriver,
   JobRunSummary,
   ListApplicationsCommand,
   ListJobRunsCommand,
+  ListJobRunsRequest,
   StartJobRunCommand,
+  StartJobRunCommandInput,
+  StartJobRunRequest,
 } from "@aws-sdk/client-emr-serverless";
 import { Globals } from "../extension";
 
@@ -60,20 +64,55 @@ export class DefaultEMRServerlessClient {
         );
         const emr = await this.createEMRServerless();
         let jobRuns: JobRun[] = [];
+        let request: ListJobRunsRequest = {
+          applicationId: applicationId,
+        };
     
         try {
-          const result = await emr.send(
-            new ListJobRunsCommand({
-              applicationId: applicationId,
-            })
-          );
-          jobRuns = result.jobRuns ?? [];
+          do {
+            const result = await emr.send(new ListJobRunsCommand(request));
+            jobRuns = jobRuns.concat(result.jobRuns ?? []);
+            if (!result.nextToken || jobRuns.length >= 100) {
+              break;
+            }
+            request['nextToken'] = result.nextToken;
+          } while (request['nextToken']);
         } catch (error) {
           vscode.window.showErrorMessage(
             "Error fetching EMR application job runs!" + error
           );
         }
-    
+
         return jobRuns;
       }
+  
+    public async startJobRun(applicationId: string, executionRoleARN: string, entryPoint: string): Promise<JobRun> {
+      this.globals.outputChannel.appendLine(
+        `EMR Serverless: Starting job run (${applicationId}).`
+      );
+
+      const emr = await this.createEMRServerless();
+      let jobRun: JobRun = {};
+
+      let jobRunParams: StartJobRunCommandInput = {
+        applicationId,
+        executionRoleArn: executionRoleARN,
+        jobDriver: {
+          sparkSubmit: {entryPoint: entryPoint}
+        }
+      };
+
+      try {
+        const result = await emr.send(
+          new StartJobRunCommand(jobRunParams)
+        );
+      } catch (error) {
+        vscode.window.showErrorMessage(
+          "There was an error running the EMR Serverless job:" + error
+        );
+      }
+
+
+      return jobRun;
+    }
 }