+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Change Log
 
+## 0.2.0 (Developer Preview) - 2022-10-10
+
+### Features
+
+- Add support for deploying single-file PySpark jobs
+- Enable the Glue Data Catalog by default for devcontainer
+
+### Bug Fixes
+
+- Paginate job runs for EMR Serverless
+
+## 0.1.0 (Developer Preview) - 2022-09-28
+
+### Features
+
+- Add ipykernel to support Jupyter notebooks
+
 ## 0.0.1 (Developer Preview) - 2022-07-27
 
 - Initial release